+++ spec/helpers/activity_helper_spec.rb
@@ -2,6 +2,9 @@ require "rails_helper"
 
 RSpec.describe ActivityHelper, type: :helper do
   let(:organisation) { create(:partner_organisation) }
+  let(:programme_activity) { create(:programme_activity) }
+  let(:project_activity) { create(:project_activity) }
+  let(:report) { create(:report) }
 
   describe "#step_is_complete_or_next?" do
     context "when the activity has passed the identification step" do
@@ -58,6 +61,132 @@ RSpec.describe ActivityHelper, type: :helper do
 
         expect(helper.link_to_activity_parent(parent: parent_activity, user: user)).to include organisation_activity_path(parent_activity.organisation, parent_activity)
         expect(helper.link_to_activity_parent(parent: parent_activity, user: user)).to include parent_activity.title
+      end
+    end
+  end
+
+  describe "#show_link_to_add_comment?" do
+    context "when activity is a programme" do
+      context "when user is authorised to add programme comments" do
+        before { authorise_creating_comments(commentable_type: :programme_activity, authorise: true) }
+
+        it "returns true" do
+          expect(helper.show_link_to_add_comment?(activity: programme_activity, report: nil)).to eq(true)
+        end
+      end
+
+      context "when user is unauthorised to add programme comments" do
+        before do
+          authorise_creating_comments(commentable_type: :programme_activity, authorise: false)
+          authorise_creating_comments(commentable_type: :project_activity, authorise: true)
+        end
+
+        it "returns false" do
+          expect(helper.show_link_to_add_comment?(activity: programme_activity, report: nil)).to eq(false)
+        end
+      end
+    end
+
+    context "when activity is a project" do
+      context "when user is authorised to add project comments" do
+        before do
+          authorise_creating_comments(commentable_type: :programme_activity, authorise: false)
+          authorise_creating_comments(commentable_type: :project_activity, authorise: true)
+        end
+
+        context "when a report exists" do
+          it "returns true" do
+            expect(helper.show_link_to_add_comment?(activity: project_activity, report: report)).to eq(true)
+          end
+        end
+
+        context "when a report does not exist" do
+          it "returns false" do
+            expect(helper.show_link_to_add_comment?(activity: project_activity, report: nil)).to eq(false)
+          end
+        end
+      end
+
+      context "when user is unauthorised to add project comments" do
+        before do
+          authorise_creating_comments(commentable_type: :programme_activity, authorise: false)
+          authorise_creating_comments(commentable_type: :project_activity, authorise: false)
+        end
+
+        context "when a report exists" do
+          it "returns false" do
+            expect(helper.show_link_to_add_comment?(activity: project_activity, report: report)).to eq(false)
+          end
+        end
+
+        context "when a report does not exist" do
+          it "returns false" do
+            expect(helper.show_link_to_add_comment?(activity: project_activity, report: nil)).to eq(false)
+          end
+        end
+      end
+    end
+  end
+
+  describe "#show_link_to_edit_comment?" do
+    context "when commentable is a programme activity" do
+      let(:existing_programme_activity_comment) { create(:comment, commentable: programme_activity, owner: create(:beis_user)) }
+
+      context "when user is authorised to edit programme comments" do
+        it "returns true" do
+          authorise_updating_comments(commentable_type: :programme_activity, authorise: true)
+
+          expect(helper.show_link_to_edit_comment?(comment: existing_programme_activity_comment)).to eq(true)
+        end
+      end
+
+      context "when user is unauthorised to add programme comments" do
+        it "returns false" do
+          authorise_updating_comments(commentable_type: :programme_activity, authorise: false)
+
+          expect(helper.show_link_to_edit_comment?(comment: existing_programme_activity_comment)).to eq(false)
+        end
+      end
+    end
+
+    context "when commentable is a project activity" do
+      let(:existing_project_activity_comment) { create(:comment, commentable: project_activity, owner: create(:partner_organisation_user), report: report) }
+
+      context "when user is authorised to add project comments" do
+        it "returns true" do
+          authorise_updating_comments(commentable_type: :project_activity, authorise: true)
+
+          expect(helper.show_link_to_edit_comment?(comment: existing_project_activity_comment)).to eq(true)
+        end
+      end
+
+      context "when user is unauthorised to add project comments" do
+        it "returns false" do
+          authorise_updating_comments(commentable_type: :project_activity, authorise: false)
+
+          expect(helper.show_link_to_edit_comment?(comment: existing_project_activity_comment)).to eq(false)
+        end
+      end
+    end
+
+    context "when commentable is not an activity" do
+      let(:actual) { create(:actual, parent_activity: project_activity) }
+      let(:existing_actual_comment) { create(:comment, commentable: actual, report: report) }
+
+      context "when user is authorised to add non-activity comments" do
+        it "returns true" do
+          authorise_updating_comments(commentable_type: :non_activity, authorise: true)
+
+          expect(helper.show_link_to_edit_comment?(comment: existing_actual_comment)).to eq(true)
+        end
+      end
+
+      context "when user is unauthorised to add non-activity comments" do
+        it "returns false" do
+          authorise_updating_comments(commentable_type: :non_activity, authorise: false)
+
+          expect(helper.show_link_to_edit_comment?(comment: existing_actual_comment)).to eq(false)
+        end
       end
     end
   end
@@ -186,6 +315,39 @@ RSpec.describe ActivityHelper, type: :helper do
 
       it "returns false" do
         expect(can_download_as_xml?(activity: activity, user: user)).to eql(false)
+      end
+    end
+  end
+
+  def authorise_creating_comments(commentable_type:, authorise: true)
+    without_partial_double_verification do
+      case commentable_type
+      when :programme_activity
+        allow(view).to receive(:policy).with(:level_b).and_return(double(:level_b_policy, create_activity_comment?: authorise))
+      when :project_activity
+        allow(view).to receive(:policy).with([:activity, :comment]).and_return(double(:activity_policy, create?: authorise))
+      end
+    end
+  end
+
+  def authorise_updating_comments(commentable_type:, authorise: true)
+    without_partial_double_verification do
+      case commentable_type
+      when :programme_activity
+        allow(view)
+          .to receive(:policy)
+          .with(:level_b)
+          .and_return(double(:level_b_policy, update_activity_comment?: authorise))
+      when :project_activity
+        allow(view)
+          .to receive(:policy)
+          .with(array_including(instance_of(Activity), instance_of(Comment)))
+          .and_return(double(:activity_policy, update?: authorise))
+      when :non_activity
+        allow(view)
+          .to receive(:policy)
+          .with(array_including(anything, instance_of(Comment)))
+          .and_return(double(:activity_policy, update?: authorise))
       end
     end
   end